Feat: feature/#6 독립 파싱 모듈 생성
적재된 raw_page 전부에서 deal을 뽑아내는 로직 수행
