+++ cherrypick-data-analysis/src/cherrypick_data_analysis/fmkorea_crawler/fmkorea_total_crawl.py
@@ -8,7 +8,7 @@ from shared.database.database import engine, Base
 import random
 from shared.query.deal_query import get_all_deal_no
 from shared.util.crawl_util import get_driver, parse_html
-from shared.dto.data_save_process import data_save_process
+from shared.save_process.data_save_process import data_save_process
 from shared.util.redis_util import *
 from time import sleep
 
@@ -21,7 +21,7 @@ def crawl_start(start_page):
     initialize_redis(Site.FMKOREA)
 
     Base.metadata.create_all(engine)
-    thread = threading.Thread(target=data_save_process, args=(q, Site.FMKOREA,)).start()
+    threading.Thread(target=data_save_process, args=(q, Site.FMKOREA,)).start()
 
 
     page = start_page
