+++ cherrypick-data-analysis/src/cherrypick_data_analysis/shared/util/crawl_util.py
@@ -1,5 +1,5 @@
 from bs4 import BeautifulSoup
-from shared.dto.page_dto import DealDTO
+from shared.save_process.page_dto import DealDTO
 import undetected_chromedriver as uc
 
 def get_driver():
